first stable working version

+++ src/main/java/ai/preferred/cerebro/index/builder/BitAndDistance.java
@@ -2,9 +2,9 @@ package ai.preferred.cerebro.index.builder;
 
 public class BitAndDistance {
     boolean bit;
-    float distance;
+    double distance;
 
-    public BitAndDistance(boolean bit, float distance) {
+    public BitAndDistance(boolean bit, double distance) {
         this.bit = bit;
         this.distance = distance;
     }
